requirements.txt
pillow
pycozmo
numpy